feat(file-service): add file service with upload and download functionalities
