+++ packages/services/src/file/file-service.interface.ts
@@ -1,0 +1,4 @@
+export interface IFileService {
+  uploadFile<T>(url: string, file: File): Promise<T>;
+  downloadFile(url: string): Promise<Blob>;
+}
