+++ packages/services/src/file/index.ts
@@ -1,0 +1,2 @@
+export * from './file-service.interface.js';
+export * from './file-service.js';
