+++ packages/services/src/http/http-service.interface.ts
@@ -4,6 +4,4 @@ export interface IHTTPService {
   patch<T>(url: string, body: unknown): Promise<T>;
   put<T>(url: string, body: unknown): Promise<T>;
   delete<T>(url: string): Promise<T>;
-  postFormData<T>(url: string, formData: FormData): Promise<T>;
-  downloadFile(url: string): Promise<Blob>;
 }
