+++ packages/services/src/index.ts
@@ -1,3 +1,4 @@
 export * from './http/index.js';
 export * from './storage/index.js';
 export * from './attachment/index.js';
+export * from './file/index.js';
